Overhaul and insanely extend v4 address parsing support.
This should allow parsing a v4 address in ANY of its possible forms,
of which there are an insane number, many of which no sane person
would ever use.  IPv4 addresses can:

- Be 4, 3, 2, or 1 dot-separated pieces, with the last piece being 8,
  16, 24, or 32 bits respectively.

- Have each piece be expressed in decimal, hex, or octal, with each
  piece being individually evaluated.


As part of this, extend our octet accumulator out to be an unsigned
long, and alter "(int)strtol()"'s to "strtoul()"'s accordingly.  And
be sure to mask off each octet with 0xff when we store it.  Do various
other rearrangements in support of this.